delete method working and redirection issue temp fix

+++ src/components/book-components/DeleteBookById.js
@@ -1,13 +1,26 @@
 import React, {useState} from "react";
+import { Navigate } from "react-router-dom";
 
 const DeleteBookById = ({deleteBookById}) => {
 
     const [idToDelete, setIdToDelete] = useState(0)
 
-    const handleIdToDeleteChange = event => setIdToDelete(event.target.value)
+    const handleIdToDeleteChange = event => setIdToDelete(event.target.value);
 
-    const handleDeleteBookById = () => {
+    const handleDeleteBookById = event => {
+        event.preventDefault();
+
+        if(!idToDelete){
+            alert("missing information")
+            return;
+        }
+
         deleteBookById(idToDelete);
+
+        setIdToDelete(0)
+
+        // to fix a redirection issue after delete form is submitted
+        return <Navigate to='/manager'/>
     }
 
     return (
